+++ src/model/user_domain.go
@@ -5,7 +5,7 @@ import (
 	"encoding/hex"
 )
 
-func NewUserDomain(email, password, name string, age int) *userDomainStruct {
+func NewUserDomain(email, password, name string, age int8) *userDomainStruct {
 	return &userDomainStruct{
 		email:    email,
 		password: password,
@@ -14,7 +14,7 @@ func NewUserDomain(email, password, name string, age int) *userDomainStruct {
 	}
 }
 
-func NewUserUpdateDomain(name string, age int) *userDomainStruct {
+func NewUserUpdateDomain(name string, age int8) *userDomainStruct {
 	return &userDomainStruct{
 		name: name,
 		age:  age,
@@ -38,6 +38,6 @@ func (ud *userDomainStruct) EncryptPass() {
 func (ud *userDomainStruct) GetEmail() string    { return ud.email }
 func (ud *userDomainStruct) GetPassword() string { return ud.password }
 func (ud *userDomainStruct) GetName() string     { return ud.name }
-func (ud *userDomainStruct) GetAge() int         { return ud.age }
+func (ud *userDomainStruct) GetAge() int8        { return ud.age }
 func (ud *userDomainStruct) GetId() string       { return ud.id }
 func (ud *userDomainStruct) SetID(id string)     { ud.id = id }
